Watchdog Timer
Microcontroller now uses a watchdog timer to implement a timeout on
receiving configuration string from the master. Once any data byte is
received at the communication port, a watchdog timer of four seconds is
initialized. In case of failure, the microcontroller is restarted. Since
now the master also implements a timeout, it ensures that in case of any
communication line distortion, thecontroller board can be initialized
with the configuration string again.